Update boolean methods, bug fix
Method 'matches' added so that the total matches between the guessed
code and the secret code can be taken into account when giving feedback
for a color that exists but is in the wrong space. This fixes a bug
where guess 'rrby' with secret code 'broo' would return '~+--' instead
of '-+--'.